backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Supabase
supabase>=2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Background jobs
arq==0.25.0
redis==5.0.1

# Data processing
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# HTTP client
httpx>=0.24.0,<0.26

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
isort==5.13.2
mypy==1.8.0
